+++ spark-load/spark-load-core/src/main/java/org/apache/doris/common/enums/JobStatus.java
@@ -1,4 +1,4 @@
-package org.apache.doris.common;
+package org.apache.doris.common.enums;
 
 public enum JobStatus {
