+++ spark-load/spark-load-core/src/main/java/org/apache/doris/load/job/Recoverable.java
@@ -1,6 +1,6 @@
 package org.apache.doris.load.job;
 
-import org.apache.doris.common.SparkLoadException;
+import org.apache.doris.exception.SparkLoadException;
 
 public interface Recoverable {
